fix long

+++ src/main/scala/xyz/cofe/json4s3/derv/FromJson.scala
@@ -148,9 +148,9 @@ object FromJson:
       case _ => Left(TypeCastFail(s"can't get int from $j"))
   given FromJson[Long] with
     def fromJson(j:AST) = j match
-      case JsInt(n) => Right(n.toShort)
-      case JsFloat(n) => Right(n.toShort)
-      case JsBig(n) => Right(n.toShort)
+      case JsInt(n) => Right(n.toLong)
+      case JsFloat(n) => Right(n.toLong)
+      case JsBig(n) => Right(n.toLong)
       case _ => Left(TypeCastFail(s"can't get long from $j"))
   given FromJson[BigInt] with
     def fromJson(j:AST) = j match
